+++ documentation/db/query.ts
@@ -1,261 +1,245 @@
 import type { MarkdownInstance } from "astro";
-import type {
-	CollectionConfig,
-	DBConfig,
-	TransformConfigToResult,
-	TransformDocumentToDocumentResult
-} from "./types";
-import { validateObjectSchema } from "./schema";
+import {
+	validateObjectSchema,
+	type ParsedSchema,
+	type SchemaObject
+} from "./schema";
 
-type MetaData = Record<string, any>;
-
-type MDFile = MarkdownInstance<{
-	_order?: number;
-	_date?: string;
-	[k: string]: any;
-}>;
-
-const metaDataFileImports = Object.entries(
-	import.meta.glob<MetaData>("../collection/**/_.json")
-).map(([fileName, importer]) => {
-	return [fileName.replace("../collection/", ""), importer] as const;
-});
-const markdownFileImports = Object.entries(
-	import.meta.glob<MDFile>("../collection/**/*.md")
-).map(([fileName, importer]) => {
-	return [fileName.replace("../collection/", ""), importer] as const;
-});
-
-type Directory = {
-	metaData: () => Promise<MetaData>;
-	directory: Record<string, Directory>;
-	document: Record<string, () => Promise<MDFile>>;
+validateObjectSchema({}, {});
+export type BaseCollectionConfig = {
+	readonly id: string;
+	readonly schema?: SchemaObject;
+	readonly docSchema?: SchemaObject;
 };
 
-const fileSystemMap = markdownFileImports.reduce(
-	(prev, curr) => {
-		const pathSections = curr[0].split("/");
-		let target: Directory = prev;
-		pathSections.forEach((section, i) => {
-			if (i === pathSections.length - 1) {
-				target.document[section.replace(".md", "")] = curr[1];
-			} else {
-				if (target.directory[section] === undefined) {
-					target.directory[section] = {
-						directory: {},
-						document: {},
-						metaData:
-							metaDataFileImports.find(([filePath]) => {
-								return (
-									filePath ===
-									`${pathSections.slice(0, i + 1).join("/")}/_.json`
-								);
-							})?.[1] ??
-							(async () => {
-								return {};
-							})
-					};
-				}
-				target = target.directory[section];
-			}
-		});
-		return prev;
-	},
+export type DBConfig = BaseCollectionConfig[];
+
+type MaybeObject<T extends {} | undefined> = T extends {} ? T : {};
+
+// create object where key is the base collection id, value is the collection config
+// and select the config using the base collection id from the query
+// validate result type extends BaseCollectionConfig
+type ExtractCollectionConfigById<
+	Id extends string,
+	Config extends DBConfig
+> = ValidateType<
 	{
-		directory: {},
-		document: {},
-		metaData: async () => {
-			return {};
-		}
-	} satisfies Directory
-) as Directory;
+		[K in Config[number] as K["id"]]: K;
+	}[Id],
+	BaseCollectionConfig
+>;
+
+type ValidateType<Target, Reference> = Target extends Reference
+	? Target
+	: never;
+
+type MDFile<Schema extends Record<string, any> = Record<string, any>> =
+	MarkdownInstance<
+		{
+			_order?: number;
+			_date?: string;
+		} & Schema
+	>;
+
+type Collection<Config extends BaseCollectionConfig = any> = {
+	_order: number;
+	metaData: ParsedSchema<MaybeObject<Config["schema"]>>;
+	children: Collection<Config>[];
+	documents: Document<Config>[];
+	id: string;
+	path: string;
+	$getAllNestedDocuments: () => Document<Config>[];
+};
+
+type Document<Config extends BaseCollectionConfig = any> = {
+	_order: number;
+	id: string;
+	metaData: ParsedSchema<MaybeObject<Config["docSchema"]>>;
+	path: string;
+	$Content: MDFile["Content"];
+	$getHeadings: MDFile["getHeadings"];
+};
+
+type CollectionMap<Config extends DBConfig = any> = {
+	[CollectionConfig in Config[number] as CollectionConfig["id"]]: Collection<CollectionConfig>;
+};
+
+const removeHiddenProperties = <T extends {}>(obj: T) => {
+	return Object.fromEntries(
+		Object.entries(obj).filter(([key]) => !key.startsWith("_"))
+	) as Omit<T, `_${string}`>;
+};
 
 export class DB<Config extends DBConfig> {
-	private config: Config;
-	public collection = <Id extends Config[number]["id"][number]>(
-		collectionId: Id
-	) => {
-		type Target = {
-			[K in Config[number] as K["id"][number]]: K;
-		}[Id];
-		const collection = this.config.find((val) => {
-			return val.id.includes(collectionId as any);
-		});
-		if (!collection) throw new Error(`Collection not found: ${collectionId}`);
-		// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-		// @ts-ignore
-		return new Collection<Target, [Id]>(collection, [collectionId]);
-	};
+	public fileMapPromise: Promise<CollectionMap>;
 	constructor(config: Config) {
-		this.config = config;
-	}
-}
-
-const transformDocumentToNode = async (
-	config: CollectionConfig,
-	path: Readonly<string[]>,
-	resolve: () => Promise<MDFile>
-): Promise<DocumentNode> => {
-	const resolvedFile = await resolve();
-	const metaDataSchema =
-		path.slice(1, -1).reduce((prev) => {
-			return prev?._ ?? null;
-		}, config as CollectionConfig | null)?.docSchema ?? {};
-	const id = path.at(-1);
-	if (id === undefined) throw new Error();
-	const metaData = validateObjectSchema(
-		metaDataSchema,
-		resolvedFile.frontmatter
-	);
-	return {
-		_type: "document",
-		_id: id,
-		_order: resolvedFile.frontmatter._order ?? -1,
-		_path: path.join("/"),
-		_Content: resolvedFile.Content,
-		_getHeadings: resolvedFile.getHeadings,
-		_baseCollectionId: path[0],
-		...Object.fromEntries(
-			Object.entries(metaData).filter(([key]) => !key.startsWith("_"))
-		)
-	};
-};
-
-class Collection<
-	Config extends CollectionConfig,
-	PathIds extends Readonly<string[]>
-> {
-	private config: Config;
-	private pathIds: PathIds;
-	public collection = <
-		Id extends Config["_"] extends {} ? Config["_"]["id"][number] : never
-	>(
-		collectionId: Id
-	) => {
-		if (!this.config._)
-			throw new Error(`Collection not found: ${collectionId}`);
-		return new Collection(
-			this.config._ as Exclude<Config["_"], undefined>,
-			[...this.pathIds, collectionId] as [...PathIds, Id]
-		);
-	};
-	public document = <Id extends string>(documentId: Id) => {
-		return new Document(
-			this.config,
-			[...this.pathIds, documentId] as const,
-			documentId
-		);
-	};
-	public get = async () => {
-		const targetDirectory = this.pathIds.reduce((prev, curr) => {
-			const dir = prev.directory[curr];
-			if (!dir) throw new Error("404");
-			return dir;
-		}, fileSystemMap);
-		const transformDirectoryToNode = async (
-			path: Readonly<string[]>,
-			dir: Directory
-		): Promise<CollectionNode> => {
-			const metaDataSchema =
-				path.slice(1).reduce((prev) => {
-					return prev?._ ?? null;
-				}, this.config as CollectionConfig | null)?.schema ?? {};
-			const resolvedMetaData = await dir.metaData();
-			const id = path.at(-1);
-			if (id === undefined) throw new Error();
-			const collections = await Promise.all(
-				Object.entries(dir.directory).map(([key, value]) =>
-					transformDirectoryToNode([...path, key], value)
+		const createFileMap = async () => {
+			const metaDataFiles = await Promise.all(
+				Object.entries(
+					import.meta.glob<{
+						default: Record<string, any>;
+					}>("../collection/**/_.json")
+				).map(async ([filePath, resolve]) => {
+					return [
+						filePath.replace("../collection/", ""),
+						await resolve()
+					] as const;
+				})
+			);
+			const markdownFiles = await Promise.all(
+				Object.entries(import.meta.glob<MDFile>("../collection/**/*.md")).map(
+					async ([filePath, resolve]) => {
+						return [
+							filePath.replace("../collection/", ""),
+							await resolve()
+						] as const;
+					}
 				)
 			);
-			const documents = await Promise.all(
-				Object.entries(dir.document).map(([key, value]) =>
-					transformDocumentToNode(this.config, [...path, key], value)
-				)
-			);
-			const sortFunction = (
-				a: CollectionNode | DocumentNode,
-				b: CollectionNode | DocumentNode
-			) => {
-				if (a._order > -1 && b._order > -1 && a._order !== b._order)
-					return a._order - b._order;
-				return a._id.localeCompare(b._id);
+			const fileMap: CollectionMap = {};
+			const getDirectory = (dirPath: string): Collection => {
+				const baseCollectionId = dirPath.split("/").at(0);
+				if (!baseCollectionId) throw new Error();
+				const collectionConfig = config.find(
+					(val) => val.id === baseCollectionId
+				);
+				if (!collectionConfig) throw new Error();
+				const isChildrenFile = (filePath: string) => {
+					const isNestedFile = filePath.startsWith(dirPath);
+					const isDirectNestedFile =
+						isNestedFile &&
+						filePath.split("/").length === dirPath.split("/").length + 1;
+					return isDirectNestedFile;
+				};
+				const metaDataFile = metaDataFiles.find(([filePath]) =>
+					isChildrenFile(filePath)
+				)?.[1] ?? {
+					default: {}
+				};
+				const order =
+					typeof metaDataFile.default._order === "number"
+						? metaDataFile.default._order
+						: -1;
+				const metaData = validateObjectSchema(
+					collectionConfig.schema ?? {},
+					removeHiddenProperties(metaDataFile.default)
+				);
+				const childrenPaths = new Set<string>();
+				for (const [filePath] of [...metaDataFiles, ...markdownFiles]) {
+					if (!filePath.startsWith(dirPath)) continue;
+					const isNestedDirectoryChildren =
+						filePath.split("/").length > dirPath.split("/").length + 1;
+					if (!isNestedDirectoryChildren) continue;
+					const childrenPath = filePath
+						.split("/")
+						.slice(0, dirPath.split("/").length + 1)
+						.join("/");
+					childrenPaths.add(childrenPath);
+				}
+				const children = [...childrenPaths]
+					.map((childPath) => {
+						const childName = childPath.split("/").at(-1);
+						if (!childName) throw new Error();
+						return getDirectory(childPath);
+					})
+					.sort(({ _order: aOrder }, { _order: bOrder }) => aOrder - bOrder);
+				const documents = markdownFiles
+					.filter(([filePath]) => isChildrenFile(filePath))
+					.map(([filePath, file]) => {
+						const fileName = filePath.split("/").at(-1);
+						if (!fileName) throw new Error();
+						const fileId = fileName.split(".").slice(0, -1).join(".");
+						const order = file.frontmatter._order ?? -1;
+						return {
+							_order: order,
+							id: fileId,
+							metaData: validateObjectSchema(
+								collectionConfig.docSchema ?? {},
+								removeHiddenProperties(file.frontmatter)
+							),
+							path: [dirPath, fileId].join("/"),
+							$Content: file.Content,
+							$getHeadings: file.getHeadings
+						} as const satisfies Document;
+					})
+					.sort(({ _order: aOrder }, { _order: bOrder }) => aOrder - bOrder);
+				const directoryId = dirPath.split("/").at(-1);
+				if (!directoryId) throw new Error();
+				const getAllNestedDocuments = (currentCollection: {
+					documents: Document<BaseCollectionConfig>[];
+					children: Collection<BaseCollectionConfig>[];
+				}): Document<Collection>[] => {
+					return [
+						...currentCollection.documents,
+						...currentCollection.children
+							.map((child) => getAllNestedDocuments(child))
+							.flat()
+					];
+				};
+				return {
+					path: dirPath,
+					id: directoryId,
+					_order: order,
+					metaData,
+					children,
+					documents,
+					$getAllNestedDocuments: () =>
+						getAllNestedDocuments({
+							documents,
+							children
+						})
+				} as const;
 			};
-			collections.sort(sortFunction);
-			documents.sort(sortFunction);
-			const metaData = validateObjectSchema(
-				metaDataSchema,
-				resolvedMetaData?.default ?? {}
-			);
-			return {
-				_type: "collection",
-				_id: id,
-				_path: path.join("/"),
-				_order: metaData._order ?? -1,
-				_collections: collections,
-				_baseCollectionId: path[0],
-				_documents: documents,
-				...Object.fromEntries(
-					Object.entries(metaData).filter(([key]) => !key.startsWith("_"))
-				)
-			};
+			for (const collectionConfig of config) {
+				fileMap[collectionConfig.id] = getDirectory(collectionConfig.id);
+			}
+			return fileMap as CollectionMap<Config>;
 		};
-		return (await transformDirectoryToNode(
-			this.pathIds,
-			targetDirectory
-		)) as TransformConfigToResult<Config, PathIds[0]>;
-	};
-	constructor(config: Config, pathIds: PathIds) {
-		this.config = config;
-		this.pathIds = pathIds;
+		this.fileMapPromise = createFileMap();
 	}
-}
-
-class Document<
-	Config extends CollectionConfig,
-	PathIds extends Readonly<string[]>,
-	Id extends string
-> {
-	private collectionConfig: Config;
-	private pathIds: PathIds;
-	private id: Id;
-	constructor(collectionConfig: Config, pathIds: PathIds, id: Id) {
-		this.collectionConfig = collectionConfig;
-		this.pathIds = pathIds;
-		this.id = id;
-	}
-	public get = async () => {
-		const targetDirectory = this.pathIds.slice(0, -1).reduce((prev, curr) => {
-			const dir = prev.directory[curr];
-			if (!dir) throw new Error("404");
-			return dir;
-		}, fileSystemMap);
-		const documentId = this.pathIds.at(-1);
-		if (documentId === undefined) throw new Error();
-		return (await transformDocumentToNode(
-			this.collectionConfig,
-			this.pathIds,
-			targetDirectory.document[documentId]
-		)) as TransformDocumentToDocumentResult<CollectionConfig, Id, PathIds[0]>;
+	public query = async <
+		PathSegments extends [Config[number]["id"], ...string[]]
+	>(
+		...pathSegments: PathSegments
+	) => {
+		const collectionMap = await this.fileMapPromise;
+		const baseCollectionId = pathSegments[0];
+		if (!(baseCollectionId in collectionMap))
+			throw new Error(`Collection "${baseCollectionId}" does not exist`);
+		const localPathSegments = pathSegments.slice(1);
+		let currentCollection = collectionMap[baseCollectionId];
+		for (const localPathSegment of localPathSegments) {
+			const collectionSearchResult =
+				currentCollection.children.find(
+					(child) => child.id === localPathSegment
+				) ?? null;
+			if (!collectionSearchResult)
+				throw new Error(`Collection ${pathSegments.join("/")} does not exist`);
+			currentCollection = collectionSearchResult;
+		}
+		type CollectionConfig = ExtractCollectionConfigById<
+			PathSegments[0],
+			Config
+		>;
+		return currentCollection as Collection<CollectionConfig>;
 	};
 }
 
-type CollectionNode = {
-	_type: "collection";
-	_id: string;
-	_path: string;
-	_order: number;
-	_baseCollectionId: string;
-	_collections: CollectionNode[];
-	_documents: DocumentNode[];
-};
+export type CollectionQuery<
+	DBInstance extends DB<any>,
+	CollectionId extends DBInstance extends DB<infer Config>
+		? Config[number]["id"]
+		: never
+> = Collection<
+	ExtractCollectionConfigById<
+		CollectionId,
+		DBInstance extends DB<infer Config> ? Config : never
+	>
+>;
 
-type DocumentNode = {
-	_type: "document";
-	_id: string;
-	_path: string;
-	_order: number;
-	_baseCollectionId: string;
-	_Content: MarkdownInstance<any>["Content"];
-	_getHeadings: MarkdownInstance<any>["getHeadings"];
-};
+export type QueryDocument<
+	DBInstance extends DB<any>,
+	CollectionId extends DBInstance extends DB<infer Config>
+		? Config[number]["id"]
+		: never
+> = CollectionQuery<DBInstance, CollectionId>["documents"][number];