+++ documentation/src/lib/db.ts
@@ -1,46 +1,26 @@
-import { DB } from "@db/query";
+import { DB, type DBConfig } from "@db/query";
 import { Optional$, String$ } from "@db/schema";
-import type { CollectionConfig, Query } from "@db/types";
 
 const config = [
 	{
-		id: ["integration"] as const,
-		_: {
-			id: ["nextjs", "sveltekit", "oauth"] as const,
-			schema: {
-				title: String$()
-			},
-			_: {
-				id: ["*"] as const,
-				schema: {
-					title: String$()
-				},
-				docSchema: {
-					title: String$()
-				}
-			}
+		id: "integration",
+		schema: {
+			title: String$()
+		},
+		docSchema: {
+			title: String$()
 		}
-	} satisfies CollectionConfig,
+	} as const,
 	{
-		id: ["main"] as const,
-		_: {
-			id: ["learn", "reference"] as const,
-			schema: {
-				title: String$()
-			},
-			_: {
-				id: ["*"] as const,
-				schema: {
-					title: String$()
-				},
-				docSchema: {
-					title: String$(),
-					redirect: Optional$(String$())
-				}
-			}
+		id: "main",
+		schema: {
+			title: String$()
+		},
+		docSchema: {
+			title: String$(),
+			redirect: Optional$(String$())
 		}
-	} satisfies CollectionConfig
-] as const;
+	} as const
+] satisfies DBConfig;
 
 export const db = new DB(config);
-export type Q = Query<typeof config>;