+++ documentation/src/lib/types.d.ts
@@ -1,1 +1,0 @@
-type ExternalIntegration = "sveltekit" | "astro" | "nextjs" | "oauth";