+++ packages/adapter-prisma/src/utils.ts
@@ -1,11 +1,20 @@
-import type { SessionSchema } from "lucia-auth";
+import type { KeySchema, SessionSchema } from "lucia-auth";
 
-// TODO: Remove bigint => number conversion on next major version
-export const convertSession = (session: SessionSchema): SessionSchema => {
-	const { active_expires, idle_expires: idleExpires, ...data } = session;
+export const convertSessionData = (
+	sessionData: SessionSchema
+): SessionSchema => {
+	const { active_expires, idle_expires: idleExpires, ...data } = sessionData;
 	return {
 		active_expires: Number(active_expires),
 		idle_expires: Number(idleExpires),
+		...data
+	};
+};
+
+export const convertKeyData = (keyData: KeySchema): KeySchema => {
+	const { expires, ...data } = keyData;
+	return {
+		expires: expires === null ? null : Number(expires),
 		...data
 	};
 };