+++ packages/adapter-kysely/CHANGELOG.md
@@ -1,8 +1,12 @@
 # CHANGELOG
 
+## 0.6.0
+
+- [Breaking] Require minimum `lucia-auth` 0.7.0
+
 ## 0.5.0
 
-- [Breaking] Require `lucia-auth` 0.6.0
+- [Breaking] Require minimum `lucia-auth` 0.6.0
 
 ## 0.4.1
 
@@ -10,7 +14,7 @@
 
 ## 0.4.0
 
-- [Breaking] Require `lucia-auth` 0.5.0
+- [Breaking] Require minimum `lucia-auth` 0.5.0
 
 ## 0.3.1
 
@@ -27,7 +31,7 @@
 
 ## 0.2.0
 
-- [Breaking] Require `lucia-auth` 0.4.0
+- [Breaking] Require minimum `lucia-auth` 0.4.0
 - [Breaking] Remove global error handler
 
 ## 0.1.1