+++ packages/adapter-prisma/CHANGELOG.md
@@ -1,12 +1,16 @@
 # CHANGELOG
 
+## 0.5.0
+
+- [Breaking] Require minimum `lucia-auth` 0.7.0
+
 ## 0.4.0
 
-- [Breaking] Require `lucia-auth` 0.6.0
+- [Breaking] Require minimum `lucia-auth` 0.6.0
 
 ## 0.3.0
 
-- [Breaking] Require `lucia-auth` 0.5.0
+- [Breaking] Require minimum `lucia-auth` 0.5.0
 
 ## 0.2.2
 
@@ -18,7 +22,7 @@
 
 ## 0.2.0
 
-- [Breaking] Require `lucia-auth` 0.4.0
+- [Breaking] Require minimum `lucia-auth` 0.4.0
 - [Breaking] Remove global error handler
 
 ## 0.1.4