+++ packages/integration-nextjs/CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## 0.5.2
+
+- Update peer dependency
+
 ## 0.5.1
 
 - Update peer dependency